+++ include/phy/sdr.hpp
@@ -68,15 +68,15 @@ inline bool has_any_except(Flags flags, Flags excluded)
 
 struct SDRConfig {
     int buffer_size = 1920;
-    double sample_rate = 1.92e6;
-    double tx_freq = 2200e6; // FDD: set rx_freq to tx_freq + 3MHz on the other node
-    double rx_freq = 2200e6; // FDD: adjust per node, see Configuration in README
+    float sample_rate = 1.92e6f;
+    float tx_freq = 2200e6f; // FDD: set rx_freq to tx_freq + 3MHz on the other node
+    float rx_freq = 2200e6f; // FDD: adjust per node, see Configuration in README
     float tx_gain = 89.0f;
     float rx_gain = 25.0f;
     bool enable_tx = true;
     bool enable_rx = true;
-    float tx_bandwidth = 1e6;
-    float rx_bandwidth = 10e6;
+    float tx_bandwidth = 1e6f;
+    float rx_bandwidth = 10e6f;
     bool init_on_start = true;
     bool exit_on_error = true;
 };
@@ -106,15 +106,15 @@ class SDR {
     int readstream(std::vector<int16_t> &send);
     int writestream(std::vector<int16_t> &send);
 
-    void set_rx_freq(double f)
+    void set_rx_freq(float f)
     {
-        logs::sdr.info("RX carrier: {:.3f} GHz", f / 1e9);
+        logs::sdr.info("RX carrier: {:.3f} GHz", f / 1e9f);
         cfg.rx_freq = f;
         flags |= Flags::APPLY_FREQUENCY;
     }
-    void set_tx_freq(double f)
+    void set_tx_freq(float f)
     {
-        logs::sdr.info("TX carrier: {:.3f} GHz", f / 1e9);
+        logs::sdr.info("TX carrier: {:.3f} GHz", f / 1e9f);
         cfg.tx_freq = f;
         flags |= Flags::APPLY_FREQUENCY;
     }
@@ -134,7 +134,7 @@ class SDR {
     bool get_wait_flag() const { return cfg.init_on_start; };
     bool get_exit_flag() const { return cfg.exit_on_error; };
     int get_buffer_size() const { return cfg.buffer_size; }
-    double get_sample_rate() const { return cfg.sample_rate; }
+    float get_sample_rate() const { return cfg.sample_rate; }
     Flags get_flags() const { return flags; }
     bool add_flag(Flags flag)
     {
